+++ packages/server/src/modules/friend/friend.service.ts
@@ -1,5 +1,5 @@
-import { AddFriendTypeEnum } from '@hello/common';
-import { Injectable } from '@nestjs/common';
+import { AddFriendTypeEnum, CustomResponseCodeEnum } from '@hello/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { FriendEntity } from './entities/friend.entity';
@@ -14,7 +14,34 @@ export class FriendService {
   async add(dto: AddFriendDTO) {
     const { userAId, userBId, userCId, addType = AddFriendTypeEnum.search } = dto;
 
-    const friendEntity = new FriendEntity();
-    friendEntity;
+    const exitA = this.friendRepository.findOne({
+      where: { user_a_id: userAId, user_b_id: userBId },
+    });
+
+    const exitB = this.friendRepository.findOne({
+      where: { user_a_id: userBId, user_b_id: userAId },
+    });
+
+    if (exitA || exitB) {
+      throw new HttpException('error', CustomResponseCodeEnum.fail);
+    }
+
+    const friendEntity1 = new FriendEntity();
+    friendEntity1.user_a_id = userAId;
+    friendEntity1.user_b_id = userBId;
+    friendEntity1.user_c_id = userCId;
+    friendEntity1.send_user_id = userAId;
+    friendEntity1.add_type = addType;
+
+    const friendEntity2: FriendEntity = {
+      ...friendEntity1,
+      user_a_id: userBId,
+      user_b_id: userAId,
+    };
+
+    await this.friendRepository.save(friendEntity1);
+    await this.friendRepository.save(friendEntity2);
+
+    return;
   }
 }